requirements.txt
streamlit>=1.28.0
google-genai>=0.7.0
